+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - A vibration effect was added that is active in the last 3 seconds when resting
+- Each workout can be assigned a colour, chosen from a predefined palette
+- An activity graph screen shows completed workouts over time, colour-coded per workout
 
 ## [1.1.0] - 2026-08-05